Renaming Basic parser into Outer.
