+++ src/javasources/KTool/buildjava.xml
@@ -66,7 +66,7 @@
 				<exclude name="**/*.java" />
 			</fileset>
 		</copy>
-		<javacc target="src/org/kframework/parser/basic/Basic.jj" javacchome="../../../lib/java"/>
+		<javacc target="src/org/kframework/parser/outer/Outer.jj" javacchome="../../../lib/java"/>
 		<javacc target="src/org/kframework/utils/kastparser/KastParser.jj" javacchome="../../../lib/java" />
 	</target>
 	<target name="clean">
