+++ src/javasources/KTool/src/org/kframework/parser/.gitignore
@@ -1,11 +1,11 @@
-basic/lib
-basic/Basic.java
-basic/BasicConstants.java
-basic/BasicTokenManager.java
-basic/ParseException.java
-basic/SimpleCharStream.java
-basic/Token.java
-basic/TokenMgrError.java
+outer/lib
+outer/Outer.java
+outer/OuterConstants.java
+outer/OuterTokenManager.java
+outer/ParseException.java
+outer/SimpleCharStream.java
+outer/Token.java
+outer/TokenMgrError.java
 concrete/lib
 !concrete/lib/string_trim_last_one_0_0.java
 !concrete/lib/string_unescape_sort_0_0.java
