+++ src/javasources/KTool/test/org/kframework/parser/outer/OuterParsingTests.java
@@ -1,5 +1,5 @@
 // Copyright (c) 2014 K Team. All Rights Reserved.
-package org.kframework.parser.basic;
+package org.kframework.parser.outer;
 
 import java.util.List;
 
@@ -12,14 +12,14 @@ import org.kframework.kil.Module;
 import org.kframework.kil.StringSentence;
 import org.kframework.kil.Syntax;
 
-public class BasicParsingTests {
+public class OuterParsingTests {
 
     @Test
     public void testLexicalRules() throws Exception {
         // TODO: remove once the new parser is fully functional
         String def = "module TEST syntax Str ::= Token{((~[\\'\\n\\r\\\\])|([\\\\]~[\\n\\r]))*} endmodule";
 
-        List<DefinitionItem> defItemList = Basic.parse("UnitTest", def, null);
+        List<DefinitionItem> defItemList = Outer.parse("UnitTest", def, null);
 
         Module mod = (Module) defItemList.get(0);
         Syntax syn = (Syntax) mod.getItems().get(0);
@@ -34,7 +34,7 @@ public class BasicParsingTests {
         String def = "module TEST rule endmodule";
 
 
-        List<DefinitionItem> defItemList = Basic.parse("UnitTest", def, null);
+        List<DefinitionItem> defItemList = Outer.parse("UnitTest", def, null);
 
         Module mod = (Module) defItemList.get(0);
         StringSentence sen = (StringSentence) mod.getItems().get(0);
